+++ soc/ite/ec/it8xxx2/CMakeLists.txt
@@ -1,4 +1,4 @@
-zephyr_sources(soc.c)
+zephyr_sources(soc.c check_regs.c vector.S)
 zephyr_include_directories(.)
 
 zephyr_library_sources_ifndef(CONFIG_RISCV_ISA_EXT_M __arithmetic.S)
