+++ soc/ite/ec/it8xxx2/check_regs.c
