+++ soc/ite/ec/it8xxx2/chip_chipregs.h
