+++ soc/ite/ec/it8xxx2/vector.S
