+++ tests/drivers/gpio/gpio_ite_it8xxx2_v2/include/chip_chipregs.h
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <../soc/ite/ec/common/chip_chipregs.h>
+#include <../soc/ite/ec/it8xxx2/chip_chipregs.h>
 
 /*
  * Macros for emulated hardware registers access.
